requirements.txt
pydub                   # Soundboard - normalise volume
googletrans             # TTS - translation
gtts                    # TTS - text to speech
discord.py[voice]       # The discord library
python-dotenv
